Move custom stylesheet to own template
