+++ core/client/layouts/master_layout/master_layout.html
@@ -1,38 +1,5 @@
 <template name="masterLayout">
-  <style>
-  .coverphoto {
-    background-color: {{ primaryColor }} !important;
-  }
-
-  .project-name-link {
-    color: {{ primaryColorText }} !important;
-  }
-
-  .introjs-bullets ul li a.active {
-    background: {{ primaryColor }} !important;
-  }
-
-  .main-navigation {
-    background-color: {{ primaryColor }} !important;
-  }
-
-  .search-box {
-    background-color: {{ primaryColor }} !important;
-  }
-
-  /* Only select top-level navigation menu items */
-  #main-navbar > ul > li > a {
-    color: {{ primaryColorText }} !important;
-  }
-
-  #site-title {
-    color: {{ primaryColorText }} !important;
-  }
-
-  #site-slogan {
-    color: {{ primaryColorText }} !important;
-  }
-  </style>
+  {{> customStylesheet }}
   {{> languageBar }}
   {{# if Template.subscriptionsReady }}
   <div class="master-layout">
